refactor: organise payment failure comms exit API
