+++ handlers/payment-failure-comms-exit-api/src/clients/index.ts
@@ -1,0 +1,1 @@
+export * from './brazeClient';
