+++ handlers/payment-failure-comms-exit-api/src/handlers/paymentFailureCommsExitController.ts
@@ -1,24 +1,11 @@
 import type { APIGatewayProxyResult } from 'aws-lambda';
-import { z } from 'zod';
 import { logger } from '@modules/logger/logger';
 import { notFound, ok } from '@modules/routing/apiGatewayResponses';
-import { defaultDeps, type RuntimeDeps } from './services';
+import { paymentFailureCommsExitResponseSchema } from '../schemas';
+import { defaultDeps } from '../services';
+import type { PaymentFailureCommsExitRequest, RuntimeDeps } from '../types';
 
-export const paymentFailureCommsExitRequestSchema = z
-	.object({
-		identityId: z.string().trim().min(1),
-	})
-	.strict();
-
-type PaymentFailureCommsExitRequest = z.infer<
-	typeof paymentFailureCommsExitRequestSchema
->;
-
-const paymentFailureCommsExitResponseSchema = z.object({
-	status: z.literal('sent'),
-});
-
-export async function paymentFailureCommsExitEndpoint(
+export async function paymentFailureCommsExitController(
 	{ identityId }: PaymentFailureCommsExitRequest,
 	deps: RuntimeDeps = defaultDeps,
 ): Promise<APIGatewayProxyResult> {
