+++ handlers/payment-failure-comms-exit-api/src/helpers/getCurrentIsoTime.ts
@@ -1,0 +1,1 @@
+export const getCurrentIsoTime = (): string => new Date().toISOString();
