+++ handlers/payment-failure-comms-exit-api/src/helpers/getEnvironmentVariable.ts
@@ -1,0 +1,4 @@
+import { getIfDefined } from '@modules/nullAndUndefined';
+
+export const getEnvironmentVariable = (name: string): string =>
+	getIfDefined(process.env[name], `${name} environment variable not set`);
