+++ handlers/payment-failure-comms-exit-api/src/schemas/appConfigSchema.ts
@@ -1,0 +1,9 @@
+import { z } from 'zod';
+
+export const appConfigSchema = z.object({
+	braze: z.object({
+		apiUrl: z.string().url(),
+		apiKey: z.string().min(1),
+		appId: z.string().min(1),
+	}),
+});
