+++ handlers/payment-failure-comms-exit-api/src/schemas/brazeTrackResponseSchema.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+
+export const brazeTrackResponseSchema = z
+	.object({
+		message: z.string().optional(),
+		errors: z.array(z.string()).optional(),
+	})
+	.passthrough();
