+++ handlers/payment-failure-comms-exit-api/src/schemas/index.ts
@@ -1,0 +1,4 @@
+export * from './appConfigSchema';
+export * from './brazeTrackResponseSchema';
+export * from './paymentFailureCommsExitRequestSchema';
+export * from './paymentFailureCommsExitResponseSchema';
