+++ handlers/payment-failure-comms-exit-api/src/schemas/paymentFailureCommsExitRequestSchema.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod';
+
+export const paymentFailureCommsExitRequestSchema = z
+	.object({
+		identityId: z.string().trim().min(1),
+	})
+	.strict();
