+++ handlers/payment-failure-comms-exit-api/src/schemas/paymentFailureCommsExitResponseSchema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const paymentFailureCommsExitResponseSchema = z.object({
+	status: z.literal('sent'),
+});
