+++ handlers/payment-failure-comms-exit-api/src/services/defaultDeps.ts
@@ -1,0 +1,10 @@
+import { getCurrentIsoTime } from '../helpers';
+import type { RuntimeDeps } from '../types';
+import { getBrazeUuidFromIdapi } from './getBrazeUuidFromIdapi';
+import { sendPaymentFailureExitEvent } from './sendPaymentFailureExitEvent';
+
+export const defaultDeps: RuntimeDeps = {
+	getBrazeUuidFromIdapi,
+	sendPaymentFailureExitEvent,
+	now: getCurrentIsoTime,
+};
