+++ handlers/payment-failure-comms-exit-api/src/types/appConfig.ts
@@ -1,0 +1,7 @@
+export type AppConfig = {
+	braze: {
+		apiUrl: string;
+		apiKey: string;
+		appId: string;
+	};
+};
