+++ handlers/payment-failure-comms-exit-api/src/types/braze.ts
@@ -1,0 +1,11 @@
+import type { paymentFailureCommsExitEventName } from '../constants';
+
+export type BrazeTrackPayload = {
+	events: Array<{
+		external_id: string;
+		app_id: string;
+		name: typeof paymentFailureCommsExitEventName;
+		time: string;
+		_update_existing_only: true;
+	}>;
+};
