+++ handlers/payment-failure-comms-exit-api/src/types/brazeTrackResponse.ts
@@ -1,0 +1,4 @@
+export type BrazeTrackResponse = {
+	message?: string;
+	errors?: string[];
+} & Record<string, unknown>;
