+++ handlers/payment-failure-comms-exit-api/src/types/paymentFailureCommsExit.ts
@@ -1,0 +1,3 @@
+export type PaymentFailureCommsExitRequest = {
+	identityId: string;
+};
