+++ handlers/payment-failure-comms-exit-api/src/types/runtimeDeps.ts
@@ -1,0 +1,8 @@
+export type RuntimeDeps = {
+	getBrazeUuidFromIdapi: (identityId: string) => Promise<string | undefined>;
+	sendPaymentFailureExitEvent: (
+		brazeUuid: string,
+		time: string,
+	) => Promise<void>;
+	now: () => string;
+};
