+++ handlers/payment-failure-comms-exit-api/test/brazeClient.test.ts
@@ -1,4 +1,4 @@
-import { BrazeClient } from '../src/brazeClient';
+import { BrazeClient } from '../src/clients';
 
 const payload = {
 	events: [
